backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
matplotlib>=3.7.0
seaborn>=0.12.0